huge redesign; launch provider itself in worker
lots of changes centering around the way the provider is started in vue. it is now a normal class started in a worker and interfaced to a single store with comlink. that means the rendering thread is not involved with any messaging and request handling anymore. the Controls component is currently the single place to control the Provider.