provider: fix WasiWorkerPool created with undefined nmax
apparently, using the public var syntax in the constructor together with inline variable assignments in the body can lead to a situation where the variable is actually undefined but the typechecker does not warn about this. when using a declaration and assigning the variable in the constructor body, the checker warns about "used before initialization" correctly. huh.