broker: avoid memory leak due to defer in loop
defer is queued at the return of a function, not the current scope (e.g. a for-loop iteration). thus the timeout contexts could never be GC'd because their cancels are kept around for the defer

ideally, you would rewrite this entire timeout logic with a resettable timer to reduce allocation pressure